+++ src/main/java/fr/maxlego08/essentials/module/ZModuleManager.java
@@ -10,6 +10,7 @@ import fr.maxlego08.essentials.module.modules.economy.EconomyModule;
 import fr.maxlego08.essentials.module.modules.hologram.HologramModule;
 import fr.maxlego08.essentials.module.modules.kit.KitModule;
 import fr.maxlego08.essentials.module.modules.chat.ChatModule;
+import fr.maxlego08.essentials.module.modules.customcommands.CustomCommandModule;
 import fr.maxlego08.essentials.module.modules.HomeModule;
 import fr.maxlego08.essentials.module.modules.ItemModule;
 import fr.maxlego08.essentials.module.modules.JoinQuitModule;
@@ -81,6 +82,7 @@ public class ZModuleManager implements ModuleManager {
         this.modules.put(StepModule.class, new StepModule(this.plugin));
         this.modules.put(AFKModule.class, new AFKModule(this.plugin));
         this.modules.put(AutoMessageModule.class, new AutoMessageModule(this.plugin));
+        this.modules.put(CustomCommandModule.class, new CustomCommandModule(this.plugin));
         if (plugin.isPaperVersion()) {
             this.modules.put(DeathMessageModule.class, new DeathMessageModule(this.plugin));
         }